# Application/Backend/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API Flux Backend",
    "version": "1.0.0",
    "description": "A comprehensive API management system for environment files, web flows, and file uploads",
    "contact": {
      "name": "API Flux Team",
      "email": "support@apiflux.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000/api",
      "description": "Development server"
    },
    {
      "url": "https://api.apiflux.com/api",
      "description": "Production server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Enter JWT token"
      }
    },
...
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/api/web-flows/{id}/hierarchy": {
      "get": {
        "summary": "Get web flow hierarchy from root to specified folder",
        "tags": [
          "WebFlows"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "The web flow ID to get hierarchy for"
          }
        ],
        "responses": {
          "200": {
            "description": "Hierarchy data retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "hierarchy": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "integer",
                            "description": "Web flow ID"
                          },
                          "name": {
                            "type": "string",
                            "description": "Web flow name"
                          },
                          "icon": {
                            "type": "string",
                            "description": "Web flow icon"
                          },
                          "isFolder": {
                            "type": "boolean",
                            "description": "Whether this is a folder"
                          },
                          "parentId": {
                            "type": "integer",
                            "nullable": true,
                            "description": "Parent web flow ID"
                          },
                          "createdAt": {
                            "type": "string",
                            "format": "date-time",
                            "description": "Creation timestamp"
                          }
                        }
                      }
                    },
                    "targetId": {
                      "type": "string",
                      "description": "The target web flow ID"
                    },
                    "totalLevels": {
                      "type": "integer",
                      "description": "Total number of levels in hierarchy"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Web flow not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Web flow not found"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/web-flows/{id}/env-files": {
      "get": {
        "summary": "Get all environment files linked to a web flow",
        "tags": [
          "WebFlows"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "The web flow ID to get linked environment files for"
          },
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "default": 1
            },
            "description": "Page number for pagination"
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "default": 10
            },
            "description": "Number of items per page"
          }
        ],
        "responses": {
          "200": {
            "description": "Linked environment files retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "envFiles": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "integer",
                            "description": "Environment file ID"
                          },
                          "name": {
                            "type": "string",
                            "description": "Environment file name"
                          },
                          "description": {
                            "type": "string",
                            "description": "Environment file description"
                          },
                          "configs": {
                            "type": "array",
                            "items": {
                              "type": "object",
                              "properties": {
                                "id": {
                                  "type": "integer"
                                },
                                "key": {
                                  "type": "string"
                                },
                                "value": {
                                  "type": "string"
                                },
                                "description": {
                                  "type": "string"
                                }
                              }
                            }
                          },
                          "userId": {
                            "type": "integer"
                          },
                          "createdAt": {
                            "type": "string",
                            "format": "date-time"
                          },
                          "updatedAt": {
                            "type": "string",
                            "format": "date-time"
                          },
                          "linkId": {
                            "type": "integer",
                            "description": "The WebFlowEnvFile link ID"
                          },
                          "linkedAt": {
                            "type": "string",
                            "format": "date-time",
                            "description": "When the file was linked to this webflow"
                          }
                        }
                      }
                    },
                    "pagination": {
                      "type": "object",
                      "properties": {
                        "page": {
                          "type": "integer"
                        },
                        "limit": {
                          "type": "integer"
                        },
                        "total": {
                          "type": "integer"
                        },
                        "pages": {
                          "type": "integer"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Web flow not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Web flow not found"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
  }
}